DEV: Do not destroy external upload stub on error in debug mode
We do not want to destroy the external upload stub records
in debug mode because they allow for investigation of problems
occuring.

+++ app/services/external_upload_manager.rb
@@ -88,7 +88,12 @@ class ExternalUploadManager
     # because at this point (calling promote_to_upload!), the multipart
     # upload would already be complete.
     Discourse.store.delete_file(external_upload_stub.key)
-    external_upload_stub.destroy!
+
+    if !SiteSetting.enable_upload_debug_mode
+      external_upload_stub.destroy!
+    else
+      external_upload_stub.update(status: ExternalUploadStub.statuses[:failed])
+    end
 
     raise
   ensure
